+++ src/main/java/br/com/java/autoflex/dto/RawMaterialRequestDTO.java
@@ -1,5 +1,7 @@
 package br.com.java.autoflex.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
@@ -16,7 +18,7 @@ public class RawMaterialRequestDTO {
 
     @NotNull(message = "quantity is required")
     @Positive(message = "quantity must be positive")
-    private Double stockQuantity;
+    private BigDecimal stockQuantity;
 
     @NotNull(message = "unit is required")
     private String unit;
